Update PiPedalModel.tsx
vite/src/pipedal/PiPedalModel.tsx - For this file only!

Expected outcome is that the server may (or may not) reboot). The purpose of the catch handler (which would happen in the case of a socket disconnect) is to make sure that an error or abort does not get reported, since it is expected.

Changes to wifi configuration, where the handler code MUST NOT PROPAGATE AN EXCEPTION.

(1) just leave this function to rot in its shameful state of decay, but do NOT relay the error.
(2) Write the audio properties (specifically) without waiting for a response. The settings are propagated to other clients (and ourselves) with a on...changed notifications from the server, so there's no actual response from the server required.

Audio properties specified:
        const audioProps = [
            "http://two-play.com/plugins/toob-player#audioFile",
            "http://two-play.com/plugins/toob-player#loop",
            "http://two-play.com/plugins/toob-player#seek",
        ];